Create wallet.js, moved class Wallet to wallet.js
